Adds an indicator in history for current step
This is the second idea from the react tutorial on
a do it yourself improvement. Your current location
in the history list is indicated by a red border (they
suggest bolding the text but I thought the border idea
was a little more evident).

This also works if you go back and forth in history.
Whatever history list item is selected corresponds
to the state of the board.