Automates square rendering
This is the third suggestion of improvements from
the react tutorial. They say
'Rewrite Board to use two loops to make the squares
instead of hardcoding them.'.

I have used nested map functions instead of for loops
because I wanted to get a better handle on using
them. But it did necessitate the creation of a
dummy array 'counter'.