Fixes bug with history list
I didn't realize before but the exact behavior I was
trying to avoid was occuring with the history list.
The first entry 'Go to game start' I didn't want a player
and move location listed because none have occured at that
point.

Correspondingly I wanted those things to be listed with
subsequent moves.